MDEV-24670: avoid OOM by linux kernel co-operative memory management
Managing memory server sizing on a congested server settings is hard.
Large database programs are the easiest memory using targets for the
OOM killer.

By using the Linux kernel memory events we can start to repond to
events and free up unused memory before we are potentially killed.

To do this we extend the storage engine handler to be able to respond to
resource events.

TODO:
* call plugin->resource_pressure() in a separate thread (for each
  plugin?)
* status variable count of the number of events.
* system server variable to control the trigger server variables
* thread caches frees
* storage engine implementation
  * innodb - MDEV-29429 ( innodb flush - helped by MDEV-25341 )
  * aria cache reductions
* Windows implementation in its connectionloop
  (https://docs.microsoft.com/en-us/windows/win32/api/memoryapi/nf-memoryapi-creatememoryresourcenotification)