MDEV-19895 Support "autoset" in SET GLOBAL for AUTO_SET system variables
Add parsing of set var = autoset; where the variable is declared with
the AUTO_SET option.

Add a auto_set boolean to the sys_var class. The point of this is the code that
uses the variable should be able to access its status and adjust based
on use. Weither that be high demand of disk requiring an innodb_log_file_size
extension or more innodb_buffer_pool_size.

Those variable declared as AUTO_SET by default are in the state
where the underlying code is permitted to change the value.

The manual setting of a system variable value will clear the auto_set
after which the underlying code using that variable should respect the
fixed value.