MDEV-21456: TCP health checks to port 3306 quickly exceed max_connect_errors
The reason for the change is TCP based health checks hit this warning
all the time, and as a result, quickly reach the max_connect_errors
and the mariadb service becomes unhealthy.

Here we remove the copying of m_handshake errors to m_connection errors
that is what cases max_connection_errors to be reached.