MDEV-30932: MSAN_OPTIONS=poison_in_dtor=1 causes failures in ~sp_lex_instr()
With sp_lex_keeper as a subclass of sp_lex_instr, the destruction
of sp_lex_keeper makes it members invalid. But the object is access
during the destructor of the parent class.

Replace the sp_lex_keeper destructor with an identical clear() function
and call the clear function in the sp_lex_instr class destructor. The clear
function resets the m_lex_resp so no double invokation is caused by
~sp_lex_keeper() calling clear().